requirements.txt
yfinance
requests
python-docx
pdfkit
docx2html
openai
